+++ spec/lib/clean_architecture/use_cases/abstract_use_case_spec.rb
@@ -1,30 +1,49 @@
 # frozen_string_literal: true
 
+require 'net/http'
 require 'clean_architecture/use_cases/abstract_use_case'
 require 'dry/monads/do'
 
 module CleanArchitecture
   module UseCases
     describe AbstractUseCase do
+      class ExampleSharedContract < Contract
+        register_macro(:check_with_an_external_system) do
+          key.failure('is on the troublemaker list') if values[key_name] == 'mike@bellroy.com'
+        end
+      end
+
+      class UnrelatedUseCase < AbstractUseCase
+        contract(ExampleSharedContract) do
+          params do
+            required(:name).filled(:str?)
+          end
+        end
+      end
+
       class ExampleUseCase < AbstractUseCase
-        params do
-          required(:email).filled(:str?)
-          required(:age).filled(:int?, gt?: 18)
+        contract(ExampleSharedContract) do
+          option :some_api
+
+          params do
+            required(:email).filled(:str?)
+            required(:age).filled(:int?, gt?: 18)
+          end
+
+          rule(:email).validate(:check_with_an_external_system)
         end
 
         include Dry::Monads::Do.for(:result)
-
-        def initialize(parameters)
-          @parameters = parameters
-        end
 
         EMAILS_OF_TROLLS = %w[
           sam@samuelgil.es
         ]
 
         def result
-          valid_params = yield result_of_validating_params(@parameters)
+          valid_params = yield result_of_validating_params
           allowed_email = yield result_of_checking_against_banned_user_list(valid_params[:email])
+
+          context(:some_api).new('http://google.com/')
 
           Dry::Monads::Success("Welcome #{allowed_email}!")
         end
@@ -39,15 +58,17 @@ module CleanArchitecture
       end
 
       subject(:result) do
-        ExampleUseCase.new(parameters).result
+        ExampleUseCase.new(params).result
       end
 
-      let(:parameters) do
-        ExampleUseCase.params.call(
+      let(:params) do
+        ExampleUseCase.parameters(
+          context: { some_api: some_api },
           email: email,
           age: age
         )
       end
+      let(:some_api) { class_double(Net::HTTP) }
 
       describe '#result_of_validating_params' do
         let(:email) { 'samuel.giles@bellroy.com' }
@@ -64,10 +85,24 @@ module CleanArchitecture
           end
         end
 
+        context 'when params are invalid based on a shared macro' do
+          let(:age) { 20 }
+          let(:email) { 'mike@bellroy.com' }
+
+          specify do
+            expect(result).to be_an_instance_of(Dry::Monads::Failure)
+            expect(result.failure).to be_an_instance_of(Errors)
+            expect(result.failure.full_messages).to eq(
+              ['Email is on the troublemaker list']
+            )
+          end
+        end
+
         context 'when params are valid' do
           let(:age) { 20 }
 
           specify do
+            expect(some_api).to receive(:new).with('http://google.com/')
             expect(result).to be_an_instance_of(Dry::Monads::Success)
             expect(result.value!).to eq "Welcome samuel.giles@bellroy.com!"
           end